Only need to implement shared_mem_cat in shared/mem.c and implement sup_tss_init inside phases/sup/tss.asm before moving on with PCI things.
Don't ask me why casting is good or bad for ACPI. Idc. if it works, it works. i mean i won't get anything for it to be correct anyway. So why would i care.

+++ kernel/phases/hdp/private/acpi/cache.h
@@ -46,7 +46,7 @@ static inline struct hdp_acpi_table_t hdp_acpi_cache_tables_xsdt(const struct hd
     }
     #endif
 
-    const uint32_t entries = (xsdt->header.length - sizeof(struct hdp_acpi_sdt_header_t)) / 8;
+    const uint64_t entries = (xsdt->header.length - sizeof(struct hdp_acpi_sdt_header_t)) / 8;
     const uint64_t* ptrs = (const uint64_t*)((const uint8_t*)xsdt + sizeof(struct hdp_acpi_sdt_header_t));
 
     struct hdp_acpi_table_t table = {0};
